Caja

+++ Entidades/Turno.cs
@@ -11,7 +11,7 @@ namespace Entidades
         public int ID_Usuario { set; get; }
         public DateTime Apertura { set; get; }
         public DateTime Cierre { set; get; }
-        public int Monto { set; get; }
+        public double Monto { set; get; }
         public int Estado { set; get; }
     }
 }
